Fix xstate/fsm standalone typechecking and slim its type-import chain
- Move the Symbol.observable global augmentation plus the lean shared types
  (EventObject, MachineContext, Values, SetupSchemas, SetupStateSchema,
  SetupStateType, ActionSchemas, GuardSchemas, InferEvents,
  InferInternalEvents) into a new base.types.ts that imports only
  schema.types.ts; the previous homes re-export them.
- fsm.ts now imports types only from base.types.ts and schema.types.ts, so an
  fsm-only program contains just 4 small package files instead of the entire
  types.ts surface.
- Add typecheck:fsm-only (CI-wired): compiles an fsm-only consumer fixture in
  isolation and fails if the program pulls in unexpected files.
