+++ packages/core/src/index.ts
@@ -195,9 +195,3 @@ export {
   terminateActor
 } from './runtimeHelpers.ts';
 export { waitFor } from './waitFor.ts';
-
-declare global {
-  interface SymbolConstructor {
-    readonly observable: symbol;
-  }
-}
